I am refactoring this entire thing.
I think I need to refactor this completely.
I want to do the same idea that what it did, but I haven't seen this
code in like 4 years or so.

I am a much better engineer now then when I first started this project
so I want to redo this project.

BREAKING CHANGE: Complete refactor